requirements.txt
aiogram==3.4.1
flask
sqlalchemy
psycopg2-binary
python-dotenv